fix(ci): line-limit gate — split oversized monitor tests and trim comments
